feat(isotp): implement ISO 15765-2 §9.8 CAN FD SF/FF escape sequences
Closes #14

Add all four missing CAN FD paths to the ISO-TP transport layer:
SF RX escape (byte 0 == 0x00 on FD frames), FF RX escape (32-bit
big-endian FF_DL when ff_dl == 0 on FD frames), SF TX escape
(is_fd + length ≤ 62), and FF TX escape (is_fd + length > 4095).

API: isotp_rx_complete_cb length and isotp_transmit length widened
from uint16_t to uint32_t. Internal rx/tx length state fields also
widened. New use_fd bool in isotp_cfg_t/isotp_ctx_t; defaults to false,
so Classic CAN callers are unaffected.

New ISOTP_FD_SF_MAX_PAYLOAD_LEN (62U) and overridable ISOTP_RX_BUF_LEN
constants. 8 new ZTEST_SUITE(test_isotp_canfd) test cases added.

+++ examples/ardep_ecu/src/main.c
@@ -643,7 +643,7 @@ static int diag_security_algo_init(void)
 
 static void on_isotp_rx_complete(
     const uint8_t *data,
-    uint16_t       length,
+    uint32_t       length,
     void          *arg)
 {
     uds_server_ctx_t *srv = (uds_server_ctx_t *)arg;
@@ -652,10 +652,10 @@ static void on_isotp_rx_complete(
     uint16_t          i;
 
     if ((data == NULL) || (srv == NULL) || (tp == NULL)) { return; }
-    if ((length == 0U) || (length > (uint16_t)UDS_MAX_PAYLOAD_LEN)) { return; }
+    if ((length == 0U) || (length > (uint32_t)UDS_MAX_PAYLOAD_LEN)) { return; }
 
-    for (i = 0U; i < length; i++) { s_req_buf.data[i] = data[i]; }
-    s_req_buf.length  = length;
+    for (i = 0U; i < (uint16_t)length; i++) { s_req_buf.data[i] = data[i]; }
+    s_req_buf.length  = (uint16_t)length;
     s_resp_buf.length = 0U;
 
     LOG_DBG("UDS RX: SID=0x%02X len=%u", (unsigned)data[0], (unsigned)length);
