+++ examples/robot_joint_controller_ecu/src/main.c
@@ -134,7 +134,7 @@ static void security_init(void)
  * ISO-TP RX completion callback
  * ---------------------------------------------------------------------------*/
 
-static void on_isotp_rx_complete(const uint8_t *data, uint16_t length, void *arg)
+static void on_isotp_rx_complete(const uint8_t *data, uint32_t length, void *arg)
 {
     uds_server_ctx_t *srv = (uds_server_ctx_t *)arg;
     isotp_ctx_t      *tp  = uds_generated_get_isotp();
@@ -142,10 +142,10 @@ static void on_isotp_rx_complete(const uint8_t *data, uint16_t length, void *arg
     uint16_t          i;
 
     if ((data == NULL) || (srv == NULL) || (tp == NULL)) { return; }
-    if ((length == 0U) || (length > (uint16_t)UDS_MAX_PAYLOAD_LEN)) { return; }
+    if ((length == 0U) || (length > (uint32_t)UDS_MAX_PAYLOAD_LEN)) { return; }
 
-    for (i = 0U; i < length; i++) { s_req_buf.data[i] = data[i]; }
-    s_req_buf.length  = length;
+    for (i = 0U; i < (uint16_t)length; i++) { s_req_buf.data[i] = data[i]; }
+    s_req_buf.length  = (uint16_t)length;
     s_resp_buf.length = 0U;
 
     (void)diag_mutex_lock(&s_session_lock);
